Cambio

+++ Laboratorio TP 3/tup2024/src/main/java/ar/edu/utn/frbb/tup/service/ClienteService.java
@@ -49,9 +49,9 @@ public class ClienteService {
     }
 
 
-    public Cliente buscarClientePorDni(long dni) {
+    public Cliente buscarClientePorDni(int dni) {
         Cliente cliente = clienteDao.find(dni, true);
-        if(cliente == null) {
+        if (cliente == null) {
             throw new IllegalArgumentException("El cliente no existe");
         }
         return cliente;
